+++ src/main/java/com/example/team3/domain/Board.java
@@ -6,9 +6,12 @@ import org.hibernate.annotations.CreationTimestamp;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -29,8 +32,11 @@ public class Board {
 	private String title;
 	private String content;
 	private String img;
-	private String writer;
 	
 	@CreationTimestamp
 	private Timestamp date;
+	
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "writer")
+	private User user;
 }
